SWAL login

+++ src/app/login/login.component.ts
@@ -2,6 +2,7 @@ import { Component, OnInit } from '@angular/core';
 import { FormControl, FormGroup, Validators } from '@angular/forms';
 import { Router } from '@angular/router';
 import { Subscription } from 'rxjs';
+import Swal from 'sweetalert2';
 import { User } from './models/User';
 import { LoginService } from './service/login.service';
 
@@ -47,26 +48,72 @@ export class LoginComponent implements OnInit {
       uPassword: this.loginForm.get('password')?.value
     }
 
-    this.loginService.userAuth(this.user).subscribe({
-      next: responseData => {
-        this.data = responseData
-        console.log(this.data.body.obj[0].uId);
-        console.log(this.data.headers.get('authorization'));
-        localStorage.setItem('token',this.data.headers.get('authorization'));
-        localStorage.setItem('userrole',this.data.headers.get('userrole'));
+    Swal.fire({
+      title: 'Please Wait',
+      allowEscapeKey: false,
+      allowOutsideClick: true,
+      background: '#fff',
+      showConfirmButton: false,
+      didOpen: () => {
+        Swal.showLoading();
+
+        this.loginService.userAuth(this.user).subscribe({
+          next: responseData => {
+            console.log(responseData);
+
+            if (responseData) {
+              Swal.fire({
+                title: 'Logged In Successfully',
+                icon: 'success',
+                showConfirmButton: false,
+                timer: 1500,
+              }).then(() => {
+                this.data = responseData
+                console.log(this.data.body.obj[0].uId);
+                console.log(this.data.headers.get('authorization'));
+                localStorage.setItem('token', this.data.headers.get('authorization'));
+                localStorage.setItem('userrole', this.data.headers.get('userrole'));
 
 
-        localStorage.setItem('userid',this.data.body.obj[0].uId);
-        localStorage.setItem('username',this.data.body.obj[0].uFName);
-        // console.log(this.data.body.obj.get('uId'));
+                localStorage.setItem('userid', this.data.body.obj[0].uId);
+                localStorage.setItem('username', this.data.body.obj[0].uFName);
+                // console.log(this.data.body.obj.get('uId'));
 
-        this.router.navigate(['/dashboard']);
-      },
-      error: err => {
-        console.log(err);
-        this.errorMessage = err.error.message;
+                this.router.navigate(['/dashboard']);
+              })
+            }
+          },
+          error: err => {
+            console.log(err);
+            Swal.fire({
+              text: err.error.message,
+              icon: 'warning'
+            })
+          }
+        })
       }
-    });
+    })
+
+    // this.loginService.userAuth(this.user).subscribe({
+    //   next: responseData => {
+    //     this.data = responseData
+    //     console.log(this.data.body.obj[0].uId);
+    //     console.log(this.data.headers.get('authorization'));
+    //     localStorage.setItem('token', this.data.headers.get('authorization'));
+    //     localStorage.setItem('userrole', this.data.headers.get('userrole'));
+
+
+    //     localStorage.setItem('userid', this.data.body.obj[0].uId);
+    //     localStorage.setItem('username', this.data.body.obj[0].uFName);
+    //     // console.log(this.data.body.obj.get('uId'));
+
+    //     this.router.navigate(['/dashboard']);
+    //   },
+    //   error: err => {
+    //     console.log(err);
+    //     this.errorMessage = err.error.message;
+    //   }
+    // });
 
 
   }
